+++ src/Config/Behavior.ts
@@ -3,7 +3,7 @@ import IBehavior from './Interfaces/IBehavior';
 class Behavior implements IBehavior {
     public showPlaceholderWhenOpen: boolean = false;
     public openOnFocus:             boolean = false;
-    public closeOnSelect:           boolean = false;
+    public closeOnSelect:           boolean = true;
     public useNativeUiOnMobile:     boolean = true;
     public loop:                    boolean = false;
     public clampMaxVisibleItems:    boolean = true;