+++ src/State/Interfaces/IActions.ts
@@ -17,6 +17,7 @@ interface IActions {
     close(): void;
     makeScrollable(): void;
     makeUnscrollable(): void;
+    startClickSelecting(): void;
     selectOption(index: number, close?: boolean): void;
     focusOption(index: number, shouldScrollToView?: boolean): void;
     search(): void;